refactor: prefer tooltip to popover
