+++ src/app/backing/components/Metrics/EstimatedRewardsMetric.tsx
@@ -2,7 +2,7 @@ import { useEstimatedRewards } from '../../hooks/useEstimatedRewards'
 import KotoQuestionMarkIcon from '@/components/Icons/KotoQuestionMarkIcon'
 import { Header } from '@/components/TypographyNew'
 import { formatCurrency } from '@/lib/utils'
-import { RifRbtcPopover } from '@/app/backing/components/Popovers/RifRbtcPopover'
+import { RifRbtcTooltip } from '@/app/backing/components/Popovers/RifRbtcTooltip'
 import { Metric, MetricTitle } from '@/components/Metric'
 import { BuildersRewards } from '@/app/collective-rewards/rewards/builders/hooks/useGetBuildersRewards'
 import { DottedUnderlineLabel } from '@/app/backing/components/DottedUnderlineLabel/DottedUnderlineLabel'
@@ -30,9 +30,11 @@ export const EstimatedRewardsMetric = ({ rewardsData }: EstimatedRewardsMetricPr
       content={
         <div className="flex flex-row items-baseline gap-2 font-rootstock-sans">
           <Header variant="h1">{formatCurrency(totalEstimatedUsd)}</Header>
-          <RifRbtcPopover totalEstimatedRbtc={totalEstimatedRbtc} totalEstimatedRif={totalEstimatedRif}>
-            <DottedUnderlineLabel className="text-lg">USD</DottedUnderlineLabel>
-          </RifRbtcPopover>
+          <RifRbtcTooltip totalEstimatedRbtc={totalEstimatedRbtc} totalEstimatedRif={totalEstimatedRif}>
+            <span>
+              <DottedUnderlineLabel className="text-lg">USD</DottedUnderlineLabel>
+            </span>
+          </RifRbtcTooltip>
         </div>
       }
     />
