+++ src/app/backing/components/RewardsInfo/RewardsInfo.tsx
@@ -6,7 +6,7 @@ import { formatCurrency } from '@/lib/utils'
 import { FC } from 'react'
 import { BackerRewardsPercentage } from '../BackerPercentage/BackerRewardsPercentage'
 import { LabeledContent } from '../LabeledContent/LabeledContent'
-import { RifRbtcPopover } from '../Popovers/RifRbtcPopover'
+import { RifRbtcTooltip } from '../Popovers/RifRbtcTooltip'
 import { DottedUnderlineLabel } from '../DottedUnderlineLabel/DottedUnderlineLabel'
 
 export interface RewardsInfoProps extends BackerRewardPercentage {
@@ -30,12 +30,14 @@ export const RewardsInfo: FC<RewardsInfoProps> = ({ current, next, estimatedRewa
                 getCombinedFiatAmount([estimatedRewards.rbtc.amount, estimatedRewards.rif.amount]),
               )}
             </Paragraph>
-            <RifRbtcPopover
+            <RifRbtcTooltip
               totalEstimatedRbtc={estimatedRewards.rbtc.amount.value}
               totalEstimatedRif={estimatedRewards.rif.amount.value}
             >
-              <DottedUnderlineLabel className="font-normal">USD</DottedUnderlineLabel>
-            </RifRbtcPopover>
+              <span>
+                <DottedUnderlineLabel className="font-normal">USD</DottedUnderlineLabel>
+              </span>
+            </RifRbtcTooltip>
           </div>
         </LabeledContent>
       )}
